fix: use marker and camera calibration in StreamingAsset
Marker descriptor and image files as well as camera calibration file
use to be in project root dir which is not taken into account when
exporting the package from Unity.
This CL reverts to 0.7.0 approach where the files in StreamingAssets
are used.
This CL also deletes the files located in project root dir to avoid
confusion and redundancy.